+++ docs/CHANGELOG.rst
@@ -5,7 +5,11 @@ Changelog
 0.6.3.2 (unreleased)
 --------------------
 
-- Nothing changed yet.
+- Add ``.../unsubscribe`` view and ``Exclude Groups`` to Mailing
+  page. When upgrading to this version, upgrade step must be run
+  on applications to add 'exclude_mailing_group' relation table
+  and add group ``unsubscribed``.
+  [matejc]
 
 
 0.6.3.1 (2014-04-18)